backend/requirements.txt
alembic
fastapi==0.100.0
uvicorn==0.23.1
sqlalchemy==2.0.21
passlib
pydantic==1.10.8
python-multipart==0.0.6
python-jose==3.3.0
psycopg2-binary==2.9.8
pytest==7.4.2
pytest-asyncio==0.21.0
httpx==0.24.1
pydantic[dotenv]
starlette==0.27.0
